角色管理

+++ src/main/java/com/fym/service/system/SystemUserService.java
@@ -1,11 +1,14 @@
 package com.fym.service.system;
 
+import com.fym.dao.system.SystemPermissionDao;
 import com.fym.dao.system.SystemRoleDao;
 import com.fym.dao.system.SystemUserDao;
 import com.fym.dao.system.SystemUserRoleDao;
 import com.fym.entity.User;
+import com.fym.entity.utils.OperEntity;
 import com.fym.entity.utils.PageEntity;
 import com.fym.entity.utils.UserManager;
+import com.fym.utils.component.OperObject;
 import com.fym.utils.data.HashPageData;
 import com.github.pagehelper.PageHelper;
 import org.slf4j.Logger;
@@ -21,11 +24,11 @@ public class SystemUserService {
     private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Resource
-    private SystemRoleDao systemRoleDao;
-    @Resource
     private SystemUserDao systemUserDao;
     @Resource
     private SystemUserRoleDao systemUserRoleDao;
+    @Resource
+    private OperObject operObject;
 
     /**
      * 获取系统用户
@@ -46,23 +49,6 @@ public class SystemUserService {
     public User getSystemUser(Object userId){
         return (userId==null)?null:systemUserDao.getUserAndRole(userId);
     }
-
-    /**
-     * 获取系统角色
-     * @param page 分页信息
-     * @return 角色
-     */
-    public List<HashPageData> getSystemRole(PageEntity page){
-        List<HashPageData> roles = null;
-        if(page!=null)
-            PageHelper.startPage(page.getPageNum(), page.getPageSize());
-        try {
-            roles = systemRoleDao.getAllRoles(page);
-        } catch (Exception e) {
-            logger.error(e.getMessage());
-        }
-        return roles;
-    }
     /**
      * 修改用户信息
      * @param manager 用户信息
@@ -75,8 +61,18 @@ public class SystemUserService {
         hash.put("id",manager.getUserId());
 
         systemUserDao.update(hash);
-        systemUserRoleDao.deleteUserRole(manager);
+        if(manager.getDelRoles().length>0)
+            systemUserRoleDao.deleteUserRole(manager);
         if(manager.getAddRoles().length>0)
             systemUserRoleDao.addRoleForUser(manager);
+    }
+
+    /**
+     * 简单操作
+     * @param oper 操作参数
+     * @param operType 操作类型
+     */
+    public void editUsers(OperEntity oper, int operType){
+        operObject.editObject(oper,operType);
     }
 }
